Fix files hirarchy, add error messages, validation
