+++ client/src/App.js
@@ -5,8 +5,8 @@ import Cart from "./pages/Cart/Cart";
 import { products } from "./fakeData";
 import { useContext, useEffect, useState } from "react";
 import { GlobalContext } from "./Contexts/GlobalContext";
-import SignIn from "./pages/SignIn/SignIn";
-import SignUp from "./pages/SignUp/SignUp";
+import SignIn from "./pages/SignInUp/SignIn";
+import SignUp from "./pages/SignInUp/SignUp";
 import Navbar from "./Components/Navbar/navbar";
 import SuppliiersPage from "./pages/suppliers/suppliersPage";
 import AddProductPage from "./pages/addProduct/addProducr";
