--- client/src/pages/SignUp/SignUp.css
@@ -1,25 +1,0 @@
-.container {
-    background-color: #ddf8f8;
-    margin: auto;
-    width: 30%;
-    padding: 20px;
-    border-radius: 5px;
-    position: absolute; 
-    left: 50%; 
-    top: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.title {
-    font-size: 30px;
-    font-weight: bold;
-    text-align: center;
-    flex: 1;
-}
-
-.form {
-    margin: 20px 0px;
-    border: 1px black solid;
-    border-radius: 5px;
-    padding: 10px;
-}
